shooting + optimazation

+++ freeShooting/Assets/Scripts/scriptebleObject/TowerScript.cs
@@ -6,9 +6,9 @@ public class TowerScript: ScriptableObject
     public new string name;
     public Sprite image;
     [SerializeField]
-    private float[] health =new float[5] ;
+    private short[] health =new short[5] ;
     [SerializeField]
-    private float[] damage = new float[5];
+    private short[] damage = new short[5];
     public GameObject prefab;
     public string description;
     
@@ -16,11 +16,11 @@ public class TowerScript: ScriptableObject
     public int level;
     
    
-    public float Get_health()
+    public short Get_health()
     {
         return health[level-1];
     }
-    public float Get_damage()
+    public short Get_damage()
     {
         return damage[level - 1];
     }
